fix: simplify rainbowkit integration
