+++ packages/debugger/app/providers.tsx
@@ -1,48 +1,31 @@
 "use client";
 
 import { RainbowKitProvider, getDefaultConfig } from "@rainbow-me/rainbowkit";
-import "@rainbow-me/rainbowkit/styles.css";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import * as React from "react";
+import { WagmiProvider } from "wagmi";
 import {
-  WagmiProvider,
-  cookieStorage,
-  cookieToInitialState,
-  createStorage,
-} from "wagmi";
-import {
-  anvil,
   arbitrum,
   base,
   mainnet,
   optimism,
   polygon,
   zora,
+  anvil,
 } from "wagmi/chains";
 
-export const config = getDefaultConfig({
+const config = getDefaultConfig({
   appName: "frames.js debugger",
   projectId: process.env.NEXT_PUBLIC_WALLETCONNECT_ID!,
   chains: [mainnet, polygon, optimism, arbitrum, base, zora, anvil],
   ssr: true,
-  storage: createStorage({
-    storage: cookieStorage,
-  }),
 });
 
 const queryClient = new QueryClient();
 
-export function Providers({
-  children,
-  cookie,
-}: {
-  cookie: string;
-  children: React.ReactNode;
-}) {
-  const initialState = cookieToInitialState(config, cookie);
-
+export function Providers({ children }: { children: React.ReactNode }) {
   return (
-    <WagmiProvider config={config} {...(initialState ? { initialState } : {})}>
+    <WagmiProvider config={config}>
       <QueryClientProvider client={queryClient}>
         <RainbowKitProvider>{children}</RainbowKitProvider>
       </QueryClientProvider>
